refactor: internal helpers & configs
