+++ src/biome/text/pipeline.py
@@ -3,46 +3,32 @@ import glob
 import inspect
 import logging
 import os
-import time
 import uuid
 from inspect import Parameter
-from threading import Thread
 from typing import Any, Dict, List, Optional, Type, Union, cast
-from urllib.error import URLError
 
 import numpy
-import uvicorn
 import yaml
 from allennlp.common import Params
-from allennlp.common.util import sanitize
-from allennlp.data import DatasetReader, Vocabulary
+from allennlp.data import Vocabulary
 from allennlp.models import load_archive
 from allennlp.models.archival import Archive
-from dask import dataframe as dd
-from dask_elk.client import DaskElasticClient
-from fastapi import FastAPI
-
-from biome.text._impl_model import AllennlpModel, _BaseModelImpl
 from biome.text.configuration import (
     PipelineConfiguration,
     TrainerConfiguration,
     VocabularyConfiguration,
 )
-from biome.text.data import DataSource
-from biome.text.errors import http_error_handling
-from biome.text.helpers import (
-    split_signature_params_by_predicate,
-    update_method_signature,
-)
-from biome.text.ui import launch_ui
-from . import constants, helpers
-from ._pipeline_helper import (
+from biome.text.helpers import update_method_signature
+from dask import dataframe as dd
+
+from . import constants
+from ._configuration import (
     ElasticsearchExplore,
     ExploreConfiguration,
     TrainConfiguration,
+    _ModelImpl,
 )
 from .model import Model
-from .modules.encoders import TimeDistributedEncoder
 from .modules.heads import TaskHead
 from .modules.heads.defs import TaskHeadSpec
 
@@ -50,18 +36,6 @@ try:
     import ujson as json
 except ModuleNotFoundError:
     import json
-
-__default_impl__ = _BaseModelImpl
-
-
-def __register(impl_class, overrides: bool = False):
-    """Register the impl. class in allennlp components registry"""
-
-    AllennlpModel.register(impl_class.__name__, exist_ok=overrides)(impl_class)
-    DatasetReader.register(impl_class.__name__, exist_ok=overrides)(impl_class)
-
-
-__register(__default_impl__, overrides=True)
 
 
 class Pipeline:
@@ -75,7 +49,7 @@ class Pipeline:
     __LOGGER = logging.getLogger(__name__)
     __TRAINING_CACHE_DATA = "instances_data"
 
-    _model: __default_impl__ = None
+    _model: _ModelImpl = None
     _config: PipelineConfiguration = None
 
     @classmethod
@@ -99,7 +73,7 @@ class Pipeline:
 
     @classmethod
     def from_config(
-        cls, config: Union[str, PipelineConfiguration], vocab_path: Optional[str] = None
+            cls, config: Union[str, PipelineConfiguration], vocab_path: Optional[str] = None
     ) -> "Pipeline":
         """Creates a pipeline from a `PipelineConfiguration` object
 
@@ -118,7 +92,7 @@ class Pipeline:
 
         if isinstance(config, str):
             config = PipelineConfiguration.from_params(Params(yaml.safe_load(config)))
-        return _EmptyPipeline(config=config, vocab=cls._vocab_from_path(vocab_path))
+        return _BlankPipeline(config=config, vocab=cls._vocab_from_path(vocab_path))
 
     @classmethod
     def from_pretrained(cls, path: str, **kwargs) -> "Pipeline":
@@ -137,15 +111,15 @@ class Pipeline:
         return _PreTrainedPipeline(pretrained_path=path, **kwargs)
 
     def train(
-        self,
-        output: str,
-        trainer: TrainerConfiguration,
-        training: str,
-        validation: Optional[str] = None,
-        test: Optional[str] = None,
-        verbose: bool = False,
-        extend_vocab: Optional[VocabularyConfiguration] = None,
-        restore: bool = True,
+            self,
+            output: str,
+            trainer: TrainerConfiguration,
+            training: str,
+            validation: Optional[str] = None,
+            test: Optional[str] = None,
+            verbose: bool = False,
+            extend_vocab: Optional[VocabularyConfiguration] = None,
+            restore: bool = True,
     ) -> None:
         """Launches a training run with the specified configurations and datasources
 
@@ -170,6 +144,7 @@ class Pipeline:
             continues training process from it
 
         """
+        from ._helpers import _allennlp_configuration
 
         self.__prepare_experiment_folder(output, restore)
         self._model.cache_data(os.path.join(output, self.__TRAINING_CACHE_DATA))
@@ -189,7 +164,7 @@ class Pipeline:
         )
 
         model.launch_experiment(
-            params=Params(_allennlp_configuration(self, config,)),
+            params=Params(_allennlp_configuration(self, config, )),
             serialization_dir=output,
         )
 
@@ -253,16 +228,16 @@ class Pipeline:
         self._model.vocab.save_to_files(path)
 
     def explore(
-        self,
-        ds_path: str,
-        explore_id: Optional[str] = None,
-        es_host: Optional[str] = None,
-        batch_size: int = 500,
-        prediction_cache_size: int = 0,
-        # TODO: do we need caching for Explore runs as well or only on serving time?
-        explain: bool = False,
-        force_delete: bool = True,
-        **metadata,
+            self,
+            ds_path: str,
+            explore_id: Optional[str] = None,
+            es_host: Optional[str] = None,
+            batch_size: int = 500,
+            prediction_cache_size: int = 0,
+            # TODO: do we need caching for Explore runs as well or only on serving time?
+            explain: bool = False,
+            force_delete: bool = True,
+            **metadata,
     ) -> dd.DataFrame:
         """Launches Explore UI for a given datasource with current model
 
@@ -292,6 +267,11 @@ class Pipeline:
             pipeline: `Pipeline`
                 A configured pipeline
         """
+        from ._helpers import (
+            _explore,
+            _show_explore,
+        )
+
         config = ExploreConfiguration(
             batch_size=batch_size,
             prediction_cache_size=prediction_cache_size,
@@ -317,6 +297,8 @@ class Pipeline:
             port: `int`
                 The port to make available the prediction service
         """
+        from ._helpers import _serve
+
         self._model = self._model.eval()
         return _serve(self, port)
 
@@ -397,7 +379,7 @@ class Pipeline:
             return None
 
     def _extend_vocab_from_sources(
-        self, vocab: Vocabulary, sources: List[str], **extra_args
+            self, vocab: Vocabulary, sources: List[str], **extra_args
     ) -> Vocabulary:
         """Extends an already created vocabulary from a list of source dictionary"""
         vocab.extend_from_instances(
@@ -425,7 +407,7 @@ class Pipeline:
             )
 
     def _load_vocabulary(
-        self, vocab_config: VocabularyConfiguration
+            self, vocab_config: VocabularyConfiguration
     ) -> Optional[Vocabulary]:
         """
         Extends a data vocabulary from a given configuration
@@ -457,10 +439,9 @@ class Pipeline:
         """Extend vocab if no vocab extension was launched before"""
         vocabulary = self._load_vocabulary(vocab_config)
         self._model.update_vocab(vocabulary)
-        self._vocab_extended = True
 
 
-class _EmptyPipeline(Pipeline):
+class _BlankPipeline(Pipeline):
     """
     Parameters
     ----------
@@ -472,16 +453,16 @@ class _EmptyPipeline(Pipeline):
     def __init__(self, config: PipelineConfiguration, **extra_args):
         self._config = config
         self._model = self.__model_from_config(self._config, **extra_args)
-        if not isinstance(self._model, __default_impl__):
+        if not isinstance(self._model, _ModelImpl):
             raise TypeError(f"Cannot load model. Wrong format of {self._model}")
         self._update_prediction_signatures()
 
     @staticmethod
     def __model_from_config(
-        config: PipelineConfiguration, **extra_params
-    ) -> __default_impl__:
+            config: PipelineConfiguration, **extra_params
+    ) -> _ModelImpl:
         """Creates a internal base model from pipeline configuration"""
-        return __default_impl__.from_params(Params({"config": config}), **extra_params)
+        return _ModelImpl.from_params(Params({"config": config}), **extra_params)
 
 
 class _PreTrainedPipeline(Pipeline):
@@ -500,15 +481,15 @@ class _PreTrainedPipeline(Pipeline):
         self._model = self.__model_from_archive(archive)
         self._config = self.__config_from_archive(archive)
 
-        if not isinstance(self._model, __default_impl__):
+        if not isinstance(self._model, _ModelImpl):
             raise TypeError(f"Cannot load model. Wrong format of {self._model}")
         self._update_prediction_signatures()
 
     @staticmethod
-    def __model_from_archive(archive: Archive) -> __default_impl__:
-        if not isinstance(archive.model, __default_impl__):
+    def __model_from_archive(archive: Archive) -> _ModelImpl:
+        if not isinstance(archive.model, _ModelImpl):
             raise ValueError(f"Wrong pipeline model: {archive.model}")
-        return cast(__default_impl__, archive.model)
+        return cast(_ModelImpl, archive.model)
 
     @staticmethod
     def __config_from_archive(archive: Archive) -> PipelineConfiguration:
@@ -519,236 +500,3 @@ class _PreTrainedPipeline(Pipeline):
     def trained_path(self) -> str:
         """Path to binary file when load from binary"""
         return self._binary
-
-
-def _serve(pipeline: Pipeline, port: int):
-    """Serves an pipeline as rest api"""
-
-    def make_app() -> FastAPI:
-        app = FastAPI()
-
-        @app.post("/predict")
-        async def predict(inputs: Dict[str, Any]):
-            with http_error_handling():
-                return sanitize(pipeline.predict(**inputs))
-
-        @app.post("/explain")
-        async def explain(inputs: Dict[str, Any]):
-            with http_error_handling():
-                return sanitize(pipeline.explain(**inputs))
-
-        @app.get("/_config")
-        async def config():
-            with http_error_handling():
-                return pipeline.config.as_dict()
-
-        @app.get("/_status")
-        async def status():
-            with http_error_handling():
-                return {"ok": True}
-
-        return app
-
-    uvicorn.run(make_app(), host="0.0.0.0", port=port)
-
-
-def _allennlp_configuration(
-    pipeline: Pipeline, config: TrainConfiguration
-) -> Dict[str, Any]:
-    """Creates a allennlp configuration for pipeline train experiment configuration"""
-
-    def trainer_configuration(trainer: TrainerConfiguration) -> Dict[str, Any]:
-        """Creates trainer configuration dict"""
-        __excluded_keys = [
-            "data_bucketing",
-            "batch_size",
-            "cache_instances",
-            "in_memory_batches",
-        ]  # Data iteration attributes
-        return {k: v for k, v in vars(trainer).items() if k not in __excluded_keys}
-
-    def iterator_configuration(
-        pipeline: Pipeline, trainer: TrainerConfiguration
-    ) -> Dict[str, Any]:
-        """Creates a data iterator configuration"""
-
-        def _forward_inputs() -> List[str]:
-            """
-            Calculate the required head.forward arguments. We use this method
-            for automatically generate data iterator sorting keys
-            """
-            required, _ = split_signature_params_by_predicate(
-                pipeline.head.forward, lambda p: p.default == inspect.Parameter.empty,
-            )
-            return [p.name for p in required] or [None]
-
-        iterator_config = {
-            "batch_size": trainer.batch_size,
-            "max_instances_in_memory": max(
-                trainer.batch_size * trainer.in_memory_batches, trainer.batch_size,
-            ),
-            "cache_instances": trainer.cache_instances,
-            "type": "basic",
-        }
-
-        if trainer.data_bucketing:
-            iterator_config.update(
-                {
-                    "sorting_keys": [
-                        [
-                            _forward_inputs()[0],
-                            "list_num_tokens"
-                            if isinstance(
-                                pipeline.model.encoder, TimeDistributedEncoder
-                            )
-                            else "num_tokens",
-                        ]
-                    ],
-                    "type": "bucket",
-                }
-            )
-
-        return iterator_config
-
-    base_config = {
-        "config": pipeline.config.as_dict(),
-        "type": __default_impl__.__name__,
-    }
-    allennlp_config = {
-        "trainer": trainer_configuration(config.trainer),
-        "iterator": iterator_configuration(pipeline, config.trainer),
-        "dataset_reader": base_config,
-        "model": base_config,
-        "train_data_path": config.training,
-        "validation_data_path": config.validation,
-        "test_data_path": config.test,
-    }
-    return copy.deepcopy({k: v for k, v in allennlp_config.items() if v})
-
-
-def _explore(
-    pipeline: Pipeline,
-    ds_path: str,
-    config: ExploreConfiguration,
-    elasticsearch: ElasticsearchExplore,
-) -> dd.DataFrame:
-    """
-    Executes a pipeline prediction over a datasource and register results int a elasticsearch index
-
-    Parameters
-    ----------
-    pipeline
-    ds_path
-    config
-    elasticsearch
-
-    Returns
-    -------
-
-    """
-    if config.prediction_cache > 0:
-        # TODO: do it
-        pipeline.init_predictions_cache(config.prediction_cache)
-
-    data_source = DataSource.from_yaml(ds_path)
-    ddf_mapped = data_source.to_mapped_dataframe()
-    # this only makes really sense when we have a predict_batch_json method implemented ...
-    n_partitions = max(1, round(len(ddf_mapped) / config.batch_size))
-
-    # a persist is necessary here, otherwise it fails for n_partitions == 1
-    # the reason is that with only 1 partition we pass on a generator to predict_batch_json
-    ddf_mapped = ddf_mapped.repartition(npartitions=n_partitions).persist()
-
-    apply_func = pipeline.explain if config.explain else pipeline.predict
-
-    ddf_mapped["annotation"] = ddf_mapped[pipeline.inputs].apply(
-        lambda x: sanitize(apply_func(**x.to_dict())), axis=1, meta=(None, object)
-    )
-
-    ddf_source = (
-        data_source.to_dataframe().repartition(npartitions=n_partitions).persist()
-    )
-    ddf_mapped["metadata"] = ddf_source.map_partitions(
-        lambda df: df.to_dict(orient="records")
-    )
-
-    # TODO @dcfidalgo we could calculate base metrics here (F1, recall & precision) using dataframe.
-    #  And include as part of explore metadata
-    #  Does it's simple???
-
-    ddf = DaskElasticClient(
-        host=elasticsearch.es_host, retry_on_timeout=True, http_compress=True
-    ).save(ddf_mapped, index=elasticsearch.es_index, doc_type=elasticsearch.es_doc)
-
-    elasticsearch.create_explore_data_index(force_delete=config.force_delete)
-    elasticsearch.create_explore_data_record(
-        {
-            **(config.metadata or {}),
-            "datasource": ds_path,
-            # TODO this should change when ui is normalized (action detail and action link naming)F
-            "explore_name": elasticsearch.es_index,
-            "model": pipeline.name,
-            "columns": ddf.columns.values.tolist(),
-            "metadata_columns": data_source.to_dataframe().columns.values.tolist(),
-            "pipeline": pipeline.type_name,
-            "output": pipeline.output,
-            "inputs": pipeline.inputs,  # backward compatibility
-            "signature": pipeline.inputs + [pipeline.output],
-            "predict_signature": pipeline.inputs,
-            "labels": pipeline.head.labels,
-            "task": pipeline.head.task_name().as_string(),
-        }
-    )
-    return ddf.persist()
-
-
-def _show_explore(elasticsearch: ElasticsearchExplore) -> None:
-    """Shows explore ui for data prediction exploration"""
-
-    def is_service_up(url: str) -> bool:
-        import urllib.request
-
-        try:
-            status_code = urllib.request.urlopen(url).getcode()
-            return 200 <= status_code < 400
-        except URLError:
-            return False
-
-    def launch_ui_app() -> Thread:
-        process = Thread(
-            target=launch_ui,
-            name="ui",
-            kwargs=dict(es_host=elasticsearch.es_host, port=ui_port),
-        )
-        process.start()
-        return process
-
-    def show_notebook_explore(url: str):
-        """Shows explore ui in a notebook cell"""
-        from IPython.core.display import HTML, display
-
-        iframe = f"<iframe src={url} width=100% height=840></iframe>"
-        display(HTML(iframe))
-
-    def show_browser_explore(url: str):
-        """Shows explore ui in a web browser"""
-        import webbrowser
-
-        webbrowser.open(url)
-
-    ui_port = 9999
-    waiting_seconds = 1
-    url = (
-        f"http://localhost:{ui_port}/projects/default/explore/{elasticsearch.es_index}"
-    )
-
-    if not is_service_up(url):
-        launch_ui_app()
-
-    time.sleep(waiting_seconds)
-    show_func = (
-        show_notebook_explore
-        if helpers.is_running_on_notebook()
-        else show_browser_explore
-    )
-    show_func(url)
